connect/pyproject.toml
[tool.poetry]
name = "ai-ready-connect"
version = "0.1.0"
description = "AI Ready Connect - M365 file event real-time detection and processing"
authors = ["GraphSuite Team"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31"
boto3 = "^1.34"
python-dotenv = "^1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
moto = {version = "^5.0", extras = ["dynamodb", "ssm", "sns", "sqs", "s3"]}
pytest-cov = "^5.0"
black = "^24.0"
flake8 = "^7.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py311"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
